cache table providers in glue catalog
