+++ crates/runtime/src/catalogconnector/glue.rs
@@ -31,12 +31,12 @@ use aws_sdk_glue::{
     error::SdkError,
     operation::{get_databases::GetDatabasesError, get_tables::GetTablesError},
 };
+use data_components::RefreshableCatalogProvider as _;
 use snafu::prelude::*;
 use std::any::Any;
 use std::sync::{Arc, LazyLock};
 
 mod provider;
-mod state;
 
 use provider::GlueCatalogProvider;
 
@@ -48,37 +48,37 @@ static VALIDATORS: LazyLock<
 
 #[derive(Debug, Snafu)]
 pub enum Error {
-    #[snafu(display("Failed to get Glue databases: {source}"))]
+    #[snafu(display("Failed to get Glue databases.\n{source}"))]
     GetDatabases { source: SdkError<GetDatabasesError> },
 
-    #[snafu(display("Failed to get Glue table from database `{database}`: {source}"))]
+    #[snafu(display("Failed to get Glue table from database `{database}`.\n{source}"))]
     GetTables {
         database: String,
         source: SdkError<GetTablesError>,
     },
 
-    #[snafu(display("Failed to build FileIO: {source}"))]
+    #[snafu(display("Failed to build FileIO.\n{source}"))]
     BuildFileIO { source: iceberg::Error },
 
-    #[snafu(display("Failed to create file input for metadata location '{location}': {source}",))]
+    #[snafu(display("Failed to create file input for metadata location '{location}'.\n{source}",))]
     CreateFileInput {
         source: iceberg::Error,
         location: String,
     },
 
-    #[snafu(display("Failed to read metadata from '{location}': {source}"))]
+    #[snafu(display("Failed to read metadata from '{location}'.\n{source}"))]
     ReadMetadata {
         source: iceberg::Error,
         location: String,
     },
 
-    #[snafu(display("Failed to deserialize metadata: {source}"))]
+    #[snafu(display("Failed to deserialize metadata.\n{source}"))]
     DeserializeMetadata { source: serde_json::Error },
 
-    #[snafu(display("Failed to build Iceberg table: {source}"))]
+    #[snafu(display("Failed to build Iceberg table.\n{source}"))]
     BuildIcebergTable { source: iceberg::Error },
 
-    #[snafu(display("Failed to create Iceberg table provider: {source}"))]
+    #[snafu(display("Failed to create Iceberg table provider.\n{source}"))]
     CreateIcebergTableProvider { source: iceberg::Error },
 
     #[snafu(display("No 'metadata_location' set on table '{table}'"))]
@@ -87,21 +87,21 @@ pub enum Error {
     #[snafu(display("No 'parameters' set on table"))]
     MissingParameters,
 
-    #[snafu(display("Parameter validation failed: {source}",))]
+    #[snafu(display("Parameter validation failed.\n{source}",))]
     ParameterValidation {
         #[snafu(source)]
         source: parameters::aws::Error,
     },
 
-    #[snafu(display("Configuration loading failed: {source}"))]
+    #[snafu(display("Configuration loading failed.\n{source}"))]
     ConfigurationLoadingFailed {
         #[snafu(source)]
         source: parameters::aws::Error,
     },
 
-    #[snafu(display("Failed to refresh catalog: {source}"))]
-    RefreshFailed {
-        #[snafu(source)]
+    #[snafu(display("Failed to create dataset `{dataset}`.\n{source}"))]
+    CreatingDataset {
+        dataset: String,
         source: Box<dyn std::error::Error + Sync + Send>,
     },
 }
@@ -134,14 +134,31 @@ impl CatalogConnector for GlueCatalog {
         runtime: Arc<Runtime>,
         catalog: &Catalog,
     ) -> super::Result<Arc<dyn data_components::RefreshableCatalogProvider>> {
-        Ok(Arc::new(
-            GlueCatalogProvider::new(self.params.clone(), catalog, runtime)
+        let app = match runtime.app.read().await.as_ref() {
+            Some(app) => Arc::clone(app),
+            None => {
+                return Err(super::Error::FailedToGetAppFromRuntime {});
+            }
+        };
+
+        let refreshable_provider = Arc::new(
+            GlueCatalogProvider::new(self.params.clone(), catalog, runtime, app)
                 .await
                 .map_err(|e| super::Error::UnableToGetCatalogProvider {
                     connector: PREFIX.to_string(),
                     connector_component: ConnectorComponent::from(catalog),
                     source: Box::new(e),
                 })?,
-        ))
+        );
+
+        refreshable_provider.refresh().await.map_err(|source| {
+            super::Error::UnableToGetCatalogProvider {
+                connector: PREFIX.to_string(),
+                connector_component: ConnectorComponent::from(catalog),
+                source,
+            }
+        })?;
+
+        Ok(refreshable_provider)
     }
 }
